Added a simple particle class
I added a few operators in the Vec class (*, *, +=) so that we can use
them to build a simple particle class. This particle class has a
function to step in time which uses Euler's method of integration, but
we will need to change this to something more robist like Verlet in the
future.