Added a Wall class
This class is basically just a wrapper around a std::vector, but it
makes things a little neater in my opinion to have simple objects
wrapped in a named class, as I have done with the Wall class.

I also added a quick test case in `src/main.py` to make sure the class
works.