Added simple particle physics
Things to change:
 - Use Iain's idea of using an LJ (or LJ-like) potential for
   inter-particle forces
 - Add walls (and their collisions)

We are very close though!

I think if we merge Iain's code with mine, we're basically there. We'll
have a rather optimized program in on time. We do need threading (but
that shouldn't be too hard: 1 thread per simulation for the genetic
algorithm I think is the best way of doing things). As for pruning (or
space-binning), in my experience its very easy to spend more time
pruning than you would have had you been not pruning, so we'll have to
see about that.